CDEC-431: Move SscMode type to lib package
